requirements.txt
markdown-it-py>=3.0.0
reportlab>=4.0.0
pygments>=2.15.0 